Updated ArtArchitecture, LineArt, Plants, Dragon, Boat, ScratchGames
Need to figure out how to make description of workshops not impact scale of photo in weird ways (and vice versa!).  Also, in the Files area, I am doing more of a brief description of what the TG contains rather than copying out materials.  Some files are missing from old workshops, so I need to work on that. Fixed URL links on these 6, but links to "download" folder are not quite working.