New page called hydraulic_bots, tagged with both E4E AND SR,, name updated in CSV and everywhere.
Declan can change the link on index to go to SR guide and change one of the links on project page to SR guide. I included links that my team used to develop this. Some other workshops also used ideas from websites, but this one especially is known without a ton of variation, so it seemed important to include. we can discuss. changed index photo to one i thought looked more dynamic.